requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.10
Django==2.2
django-ckeditor==6.1.0
django-environ==0.7.0
django-jet==1.0.8
django-js-asset==1.2.2
django-parler==2.3
django-rosetta==0.9.8
idna==3.3
Pillow==8.3.2
polib==1.1.1
pycodestyle==2.7.0
pytz==2021.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
